^ fonts: inconsolata: revert to old ver, bug makes it broken with Xft
reverts commit d95f070c

there is a bug in later Inconsolata (v3+?) where it has extra wide
spacing.  this is apparently due to a bug in Xft, which will never be
fixed (?)

follow https://github.com/googlefonts/Inconsolata/issues/42 for when
this will work.  the mentioned patch for 'st' does work btw, but urxvt
for example still fails to load, we don't know what else will be
affected by this, so we will just revert until this is either fixed in
Xft, or fixed by a release of Inconsolata without ligatures, as
described in https://github.com/googlefonts/Inconsolata/issues/35