+ fonts: inconsolata3: v3 variable font renamed as inconsolata3
this puts the newer inconsolata back (see revert a few commits back),
but renamed to something we can opt-in to, so older apps will continue
to work on the older version of the font and we can keep both and select
which one.

to rename it, used a font rename tool, since, apparently, one cannot
just rename the file, that doesn't work.

once the xft fix or some workaround with st implemented (see
https://github.com/googlefonts/Inconsolata/issues/ number 42), we will
replace the v2 font.  probably.  maybe.